Add workaround for scalar P in Meteorology
When the Meteorology component is given a scalar precipitation
rate, P, it's silently zeroed out after every update. In order
to provide a more pleasant user experience, if a scalar P value
is set, I convert it to a time series by replicating the value
over the number of time steps in the simulation.

This may cause problems when coupling components (we don't know).